Did some cleanup around the exams and the exam list view
